Fix remote test node id generation
Ids are normally generated by the IP/CH. When using a remote connection,
as the IP uses asynchronous communication, and thus the result is not
returned to the caller, it is hard to acquire the generated node's id.

The workaround here is that a node-id is preset by the caller, but this
is only necessary and allowed for testing.