Remove RemoteInfraProcessor feature
It was an architectural mistake. The Enactor and the InfraProcessor are
tightly coupled: 1 infrastructure is associated with a (1-1) pair of an
Enactor and an InfraProcessor.

Asynchrounous communication cuts the backward error channel: the Enactor
cannot be notified of an error affecting the whole infrastructure. As
the Enactor is lightweight, the original reason to decouple it in this
manner--it can be discarded after use to free resources--is void. On the
other hand, the asynchronity creates race conditions, needs an RPC
management channel, and cuts all flow of information backward (incl.
exceptions).

Synchronous RPC calls would work, but they bear no advantages as there
is always exactly *one* IP at the end of the queue. Distribution, load
balancing, message routing are not needed. Splitting the Enactor and the
IP over two hosts is also useless, as the Enactor has virtually no
resource requirements. Also, these components turned out to be tightly
coupled.

As a resolution, the support for remote communication between the
Enactor and the InfraProcessor are hereby dropped. The Enactor will
expect, accept, and use return values and exceptions from IP Commands.

The functional scopes of the two components are maintained, or rather
reinforced:
- The InfraProcessor makes node-level decisions in the scope of the
  command.
- The Enactor makes infrastructure-level decisions (e.g. suspending the
  infrastructure upon critical errors.

OCD-83 #comment Remote communication feature dropped; this issue became a whole lot simpler.

OCD-53 OCD-123 #close Feature dropped