Amend NodeCreationErrors if needed
Some backend code may not be able to access the instance_data that is
being built (esp. if the code takes part in building it;
synchronization, for example). In such cases if a NodeCreationError
needs to be raised, its instance_data cannot be initialized; but it can
be amended properly at this point in the code.

OCD-125